# docs/swagger.yaml
definitions:
  rv_internal_domain_dto_request.ConfimationCodeRequest:
    properties:
      code:
        type: string
      email:
        type: string
      newPassword:
        type: string
    required:
    - code
    - email
    type: object
  rv_internal_domain_dto_request.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    type: object
  rv_internal_domain_dto_request.RegisterCredentials:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  rv_internal_domain_dto_response.RegisterResponse:
    properties:
      userId:
        type: string
    type: object
  rv_internal_domain_dto_response.SendCodeResponse:
    properties:
      nextCodeDelay:
        $ref: '#/definitions/time.Duration'
    type: object
  rv_internal_domain_services_token.UserTokens:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    required:
    - accessToken
    - refreshToken
    type: object
  rv_pkg_response.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/rv_pkg_response.meta'
      pagination:
        $ref: '#/definitions/rv_pkg_response.pagination'
    type: object
  rv_pkg_response.meta:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
      requestId:
        type: string
    type: object
  rv_pkg_response.pagination:
    properties:
      page:
        type: integer
      pages:
        type: integer
      perPage:
        type: integer
    type: object
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
info:
  contact: {}
  description: This is reel views api service.
  title: REEL VIEWS API
  version: "1.0"
paths:
  /rl/api/v1/auth/code:
    post:
      description: register new user
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/rv_internal_domain_dto_request.LoginRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/rv_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/rv_internal_domain_dto_response.SendCodeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rv_pkg_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rv_pkg_response.Response'
      summary: send_confirm_code
      tags:
      - auth
  /rl/api/v1/auth/confirm:
    post:
      description: Подтверждение кода для подтверждения почты, либо сброса пароля.
        Если сброс пароля, то newPassword обязательное поле.
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/rv_internal_domain_dto_request.ConfimationCodeRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rv_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rv_pkg_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rv_pkg_response.Response'
      summary: confirm_code
      tags:
      - auth
  /rl/api/v1/auth/forgot:
    post:
      description: Сброс пароля
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/rv_internal_domain_dto_request.LoginRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/rv_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/rv_internal_domain_dto_response.SendCodeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rv_pkg_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rv_pkg_response.Response'
      summary: forgot_password
      tags:
      - auth
  /rl/api/v1/auth/login:
    post:
      description: Получение access,refresh токенов по почте и паролю
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/rv_internal_domain_dto_request.LoginRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/rv_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/rv_internal_domain_services_token.UserTokens'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rv_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rv_pkg_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rv_pkg_response.Response'
      summary: login
      tags:
      - auth
  /rl/api/v1/auth/refresh:
    post:
      description: Получение access,refresh токенов по access, refresh токенам
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/rv_internal_domain_services_token.UserTokens'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/rv_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/rv_internal_domain_services_token.UserTokens'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rv_pkg_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rv_pkg_response.Response'
      summary: refresh_tokens
      tags:
      - auth
  /rl/api/v1/auth/register:
    post:
      description: register new user
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/rv_internal_domain_dto_request.RegisterCredentials'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/rv_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/rv_internal_domain_dto_response.RegisterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rv_pkg_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rv_pkg_response.Response'
      summary: register_user
      tags:
      - auth
swagger: "2.0"